move _sklearn_compat to externals
