+++ skore/src/skore/externals/_skrub_compat.py
